Extract bandwidth actions to separate header
