+++ MANIFEST
@@ -43,6 +43,7 @@ include/hbwmalloc.h
 include/memkind.h
 include/memkind/internal/heap_manager.h
 include/memkind/internal/memkind_arena.h
+include/memkind/internal/memkind_bandwidth.h
 include/memkind/internal/memkind_default.h
 include/memkind/internal/memkind_gbtlb.h
 include/memkind/internal/memkind_hbw.h
@@ -356,6 +357,7 @@ src/heap_manager.c
 src/memkind-hbw-nodes.c
 src/memkind.c
 src/memkind_arena.c
+src/memkind_bandwidth.c
 src/memkind_default.c
 src/memkind_gbtlb.c
 src/memkind_hbw.c
