+++ Makefile.am
@@ -31,6 +31,7 @@ libmemkind_la_SOURCES = src/hbwmalloc.c \
                         src/heap_manager.c \
                         src/memkind.c \
                         src/memkind_arena.c \
+                        src/memkind_bandwidth.c \
                         src/memkind_default.c \
                         src/memkind_gbtlb.c \
                         src/memkind_hbw.c \
@@ -54,6 +55,7 @@ include_HEADERS = include/hbw_allocator.h \
 
 noinst_HEADERS =  include/memkind/internal/heap_manager.h \
                   include/memkind/internal/memkind_arena.h \
+                  include/memkind/internal/memkind_bandwidth.h \
                   include/memkind/internal/memkind_default.h \
                   include/memkind/internal/memkind_gbtlb.h \
                   include/memkind/internal/memkind_hbw.h \
