+++ src/memkind_hbw.c
@@ -25,6 +25,7 @@
 #include <memkind/internal/memkind_hbw.h>
 #include <memkind/internal/memkind_default.h>
 #include <memkind/internal/memkind_hugetlb.h>
+#include <memkind/internal/memkind_bandwidth.h>
 #include <memkind/internal/memkind_arena.h>
 #include <memkind/internal/memkind_private.h>
 #include <memkind/internal/memkind_log.h>
@@ -45,10 +46,6 @@
 #include <sched.h>
 #include <stdint.h>
 #include <assert.h>
-
-#define BANDWIDTH_NODE_HIGH_VAL       2
-#define BANDWIDTH_NODE_LOW_VAL        1
-#define BANDWIDTH_NODE_NOT_PRESENT    0
 
 MEMKIND_EXPORT struct memkind_ops MEMKIND_HBW_OPS = {
     .create = memkind_arena_create,
@@ -184,49 +181,21 @@ MEMKIND_EXPORT struct memkind_ops MEMKIND_HBW_INTERLEAVE_OPS = {
     .finalize = memkind_arena_finalize
 };
 
-struct numanode_bandwidth_t {
-    int numanode;
-    int bandwidth;
-};
-
-struct bandwidth_nodes_t {
-    int bandwidth;
-    int num_numanodes;
-    int *numanodes;
-};
-
-struct memkind_hbw_closest_numanode_t {
+struct memkind_closest_numanode_t {
     int init_err;
     int num_cpu;
     int *closest_numanode;
 };
 
-static struct memkind_hbw_closest_numanode_t memkind_hbw_closest_numanode_g;
+static struct memkind_closest_numanode_t memkind_hbw_closest_numanode_g;
 static pthread_once_t memkind_hbw_closest_numanode_once_g = PTHREAD_ONCE_INIT;
 
 static void memkind_hbw_closest_numanode_init(void);
 
-static int create_bandwidth_nodes(int num_bandwidth, const int *bandwidth,
-                                  int *num_unique, struct bandwidth_nodes_t **bandwidth_nodes);
-
-static int set_closest_numanode(int num_unique,
-                                const struct bandwidth_nodes_t *bandwidth_nodes,
-                                int num_cpunode, int *closest_numanode);
-
-static int numanode_bandwidth_compare(const void *a, const void *b);
-
 // This declaration is necessary, cause it's missing in headers from libnuma 2.0.8
 extern unsigned int numa_bitmask_weight(const struct bitmask *bmp );
 
-static void assign_arbitrary_bandwidth_values(int *bandwidth,
-                                              struct bitmask *hbw_nodes);
-
 static int fill_bandwidth_values_heuristically (int *bandwidth);
-
-static int fill_bandwidth_values_from_enviroment(int *bandwidth,
-                                                 char *hbw_nodes_env);
-
-static int fill_nodes_bandwidth(int *bandwidth);
 
 MEMKIND_EXPORT int memkind_hbw_check_available(struct memkind *kind)
 {
@@ -248,7 +217,7 @@ MEMKIND_EXPORT int memkind_hbw_get_mbind_nodemask(struct memkind *kind,
 {
     int cpu;
     struct bitmask nodemask_bm = {maxnode, nodemask};
-    struct memkind_hbw_closest_numanode_t *g =
+    struct memkind_closest_numanode_t *g =
             &memkind_hbw_closest_numanode_g;
     pthread_once(&memkind_hbw_closest_numanode_once_g,
                  memkind_hbw_closest_numanode_init);
@@ -270,7 +239,7 @@ MEMKIND_EXPORT int memkind_hbw_all_get_mbind_nodemask(struct memkind *kind,
 {
     int cpu;
     struct bitmask nodemask_bm = {maxnode, nodemask};
-    struct memkind_hbw_closest_numanode_t *g =
+    struct memkind_closest_numanode_t *g =
             &memkind_hbw_closest_numanode_g;
     pthread_once(&memkind_hbw_closest_numanode_once_g,
                  memkind_hbw_closest_numanode_init);
@@ -282,23 +251,6 @@ MEMKIND_EXPORT int memkind_hbw_all_get_mbind_nodemask(struct memkind *kind,
         }
     }
     return g->init_err;
-}
-
-static void assign_arbitrary_bandwidth_values(int *bandwidth,
-                                              struct bitmask *hbw_nodes)
-{
-    int i, nodes_num = numa_num_configured_nodes();
-
-    for (i=0; i<NUMA_NUM_NODES; i++) {
-        if (i >= nodes_num) {
-            bandwidth[i] = BANDWIDTH_NODE_NOT_PRESENT;
-        } else if (numa_bitmask_isbitset(hbw_nodes, i)) {
-            bandwidth[i] = BANDWIDTH_NODE_HIGH_VAL;
-        } else {
-            bandwidth[i] = BANDWIDTH_NODE_LOW_VAL;
-        }
-    }
-
 }
 
 typedef struct registers_t {
@@ -384,18 +336,6 @@ static int get_high_bandwidth_nodes(struct bitmask *hbw_node_mask)
     return MEMKIND_ERROR_UNAVAILABLE;
 }
 
-static int fill_bandwidth(int *bandwidth)
-{
-    struct bitmask *hbw_node_mask = numa_allocate_nodemask();
-    int ret = get_high_bandwidth_nodes(hbw_node_mask);
-    if(ret == 0) {
-        assign_arbitrary_bandwidth_values(bandwidth, hbw_node_mask);
-    }
-    numa_bitmask_free(hbw_node_mask);
-
-    return ret;
-}
-
 ///This function tries to fill bandwidth array based on knowledge about known CPU models
 static int fill_bandwidth_values_heuristically(int *bandwidth)
 {
@@ -409,7 +349,7 @@ static int fill_bandwidth_values_heuristically(int *bandwidth)
     switch(cpu.model) {
         case CPU_MODEL_KNL:
         case CPU_MODEL_KNM: {
-            int ret = fill_bandwidth(bandwidth);
+            int ret = bandwidth_fill(bandwidth, get_high_bandwidth_nodes);
             if(ret == 0) {
                 log_info("Detected High Bandwidth Memory.");
             }
@@ -419,43 +359,12 @@ static int fill_bandwidth_values_heuristically(int *bandwidth)
             return MEMKIND_ERROR_UNAVAILABLE;
     }
 }
-
-static int fill_bandwidth_values_from_enviroment(int *bandwidth,
-                                                 char *hbw_nodes_env)
-{
-    struct bitmask *hbw_nodes_bm = numa_parse_nodestring(hbw_nodes_env);
-
-    if (!hbw_nodes_bm) {
-        log_err("Wrong MEMKIND_HBW_NODES environment value.");
-        return MEMKIND_ERROR_ENVIRON;
-    } else {
-        assign_arbitrary_bandwidth_values(bandwidth, hbw_nodes_bm);
-        numa_bitmask_free(hbw_nodes_bm);
-    }
-
-    return 0;
-}
-
-static int fill_nodes_bandwidth(int *bandwidth)
-{
-    char *hbw_nodes_env;
-
-    hbw_nodes_env = getenv("MEMKIND_HBW_NODES");
-    if (hbw_nodes_env) {
-        log_info("Environment variable MEMKIND_HBW_NODES detected: %s.", hbw_nodes_env);
-        return fill_bandwidth_values_from_enviroment(bandwidth, hbw_nodes_env);
-    }
-
-    return fill_bandwidth_values_heuristically(bandwidth);
-}
-
 static void memkind_hbw_closest_numanode_init(void)
 {
-    struct memkind_hbw_closest_numanode_t *g =
+    struct memkind_closest_numanode_t *g =
             &memkind_hbw_closest_numanode_g;
     int *bandwidth = NULL;
     int num_unique = 0;
-    int i;
 
     struct bandwidth_nodes_t *bandwidth_nodes = NULL;
 
@@ -469,22 +378,18 @@ static void memkind_hbw_closest_numanode_init(void)
         goto exit;
     }
 
-    g->init_err = fill_nodes_bandwidth(bandwidth);
+    g->init_err = bandwidth_fill_nodes(bandwidth,
+                                       fill_bandwidth_values_heuristically, "MEMKIND_HBW_NODES");
     if (g->init_err)
         goto exit;
 
-    g->init_err = create_bandwidth_nodes(NUMA_NUM_NODES, bandwidth,
+    g->init_err = bandwidth_create_nodes(NUMA_NUM_NODES, bandwidth,
                                          &num_unique, &bandwidth_nodes);
     if (g->init_err)
         goto exit;
 
-    g->init_err = set_closest_numanode(num_unique, bandwidth_nodes,
-                                       g->num_cpu, g->closest_numanode);
-
-    for(i=0; i<bandwidth_nodes[num_unique-1].num_numanodes; i++) {
-        log_info("NUMA node %d is high-bandwidth memory.",
-                 bandwidth_nodes[num_unique-1].numanodes[i]);
-    }
+    g->init_err = bandwidth_set_closest_numanode(num_unique, bandwidth_nodes,
+                                                 g->num_cpu, g->closest_numanode);
 
 exit:
 
@@ -495,178 +400,6 @@ exit:
         jemk_free(g->closest_numanode);
         g->closest_numanode = NULL;
     }
-}
-
-
-static int create_bandwidth_nodes(int num_bandwidth, const int *bandwidth,
-                                  int *num_unique, struct bandwidth_nodes_t **bandwidth_nodes)
-{
-    /***************************************************************************
-    *   num_bandwidth (IN):                                                    *
-    *       number of numa nodes and length of bandwidth vector.               *
-    *   bandwidth (IN):                                                        *
-    *       A vector of length num_bandwidth that gives bandwidth for          *
-    *       each numa node, zero if numa node has unknown bandwidth.           *
-    *   num_unique (OUT):                                                      *
-    *       number of unique non-zero bandwidth values in bandwidth            *
-    *       vector.                                                            *
-    *   bandwidth_nodes (OUT):                                                 *
-    *       A list of length num_unique sorted by bandwidth value where        *
-    *       each element gives a list of the numa nodes that have the          *
-    *       given bandwidth.                                                   *
-    *   RETURNS zero on success, error code on failure                         *
-    ***************************************************************************/
-    int err = 0;
-    int i, j, k, l, last_bandwidth;
-    struct numanode_bandwidth_t *numanode_bandwidth = NULL;
-    *bandwidth_nodes = NULL;
-    /* allocate space for sorting array */
-    numanode_bandwidth = jemk_malloc(sizeof(struct numanode_bandwidth_t) *
-                                     num_bandwidth);
-    if (!numanode_bandwidth) {
-        err = MEMKIND_ERROR_MALLOC;
-        log_err("jemk_malloc() failed.");
-    }
-    if (!err) {
-        /* set sorting array */
-        j = 0;
-        for (i = 0; i < num_bandwidth; ++i) {
-            if (bandwidth[i] != BANDWIDTH_NODE_NOT_PRESENT) {
-                numanode_bandwidth[j].numanode = i;
-                numanode_bandwidth[j].bandwidth = bandwidth[i];
-                ++j;
-            }
-        }
-        /* ignore zero bandwidths */
-        num_bandwidth = j;
-        if (num_bandwidth == 0) {
-            err = MEMKIND_ERROR_UNAVAILABLE;
-        }
-    }
-    if (!err) {
-        qsort(numanode_bandwidth, num_bandwidth,
-              sizeof(struct numanode_bandwidth_t), numanode_bandwidth_compare);
-        /* calculate the number of unique bandwidths */
-        *num_unique = 1;
-        last_bandwidth = numanode_bandwidth[0].bandwidth;
-        for (i = 1; i < num_bandwidth; ++i) {
-            if (numanode_bandwidth[i].bandwidth != last_bandwidth) {
-                last_bandwidth = numanode_bandwidth[i].bandwidth;
-                ++*num_unique;
-            }
-        }
-        /* allocate output array */
-        *bandwidth_nodes = (struct bandwidth_nodes_t *)jemk_malloc(
-                               sizeof(struct bandwidth_nodes_t) * (*num_unique) +
-                               sizeof(int) * num_bandwidth);
-        if (!*bandwidth_nodes) {
-            err = MEMKIND_ERROR_MALLOC;
-            log_err("jemk_malloc() failed.");
-        }
-    }
-    if (!err) {
-        /* populate output */
-        (*bandwidth_nodes)[0].numanodes = (int *)(*bandwidth_nodes + *num_unique);
-        last_bandwidth = numanode_bandwidth[0].bandwidth;
-        k = 0;
-        l = 0;
-        for (i = 0; i < num_bandwidth; ++i, ++l) {
-            (*bandwidth_nodes)[0].numanodes[i] = numanode_bandwidth[i].numanode;
-            if (numanode_bandwidth[i].bandwidth != last_bandwidth) {
-                (*bandwidth_nodes)[k].num_numanodes = l;
-                (*bandwidth_nodes)[k].bandwidth = last_bandwidth;
-                l = 0;
-                ++k;
-                (*bandwidth_nodes)[k].numanodes = (*bandwidth_nodes)[0].numanodes + i;
-                last_bandwidth = numanode_bandwidth[i].bandwidth;
-            }
-        }
-        (*bandwidth_nodes)[k].num_numanodes = l;
-        (*bandwidth_nodes)[k].bandwidth = last_bandwidth;
-    }
-    if (numanode_bandwidth) {
-        jemk_free(numanode_bandwidth);
-    }
-    if (err) {
-        if (*bandwidth_nodes) {
-            jemk_free(*bandwidth_nodes);
-        }
-    }
-    return err;
-}
-
-static int set_closest_numanode(int num_unique,
-                                const struct bandwidth_nodes_t *bandwidth_nodes,
-                                int num_cpunode, int *closest_numanode)
-{
-    /***************************************************************************
-    *   num_unique (IN):                                                       *
-    *       Length of bandwidth_nodes vector.                                  *
-    *   bandwidth_nodes (IN):                                                  *
-    *       Output vector from create_bandwitdth_nodes().                      *
-    *   num_cpunode (IN):                                                      *
-    *       Number of cpu's and length of closest_numanode.                    *
-    *   closest_numanode (OUT):                                                *
-    *       Vector that maps cpu index to closest numa node of the specified   *
-    *       bandwidth.                                                         *
-    *   RETURNS zero on success, error code on failure                         *
-    ***************************************************************************/
-    int err = 0;
-    int min_distance, distance, i, j, old_errno, min_unique;
-    struct bandwidth_nodes_t match;
-    match.bandwidth = -1;
-    int target_bandwidth = bandwidth_nodes[num_unique-1].bandwidth;
-
-    for (i = 0; i < num_cpunode; ++i) {
-        closest_numanode[i] = -1;
-    }
-    for (i = 0; i < num_unique; ++i) {
-        if (bandwidth_nodes[i].bandwidth == target_bandwidth) {
-            match = bandwidth_nodes[i];
-            break;
-        }
-    }
-    if (match.bandwidth == -1) {
-        err = MEMKIND_ERROR_UNAVAILABLE;
-    } else {
-        for (i = 0; i < num_cpunode; ++i) {
-            min_distance = INT_MAX;
-            min_unique = 1;
-            for (j = 0; j < match.num_numanodes; ++j) {
-                old_errno = errno;
-                distance = numa_distance(numa_node_of_cpu(i),
-                                         match.numanodes[j]);
-                errno = old_errno;
-                if (distance < min_distance) {
-                    min_distance = distance;
-                    closest_numanode[i] = match.numanodes[j];
-                    min_unique = 1;
-                } else if (distance == min_distance) {
-                    min_unique = 0;
-                }
-            }
-            if (!min_unique) {
-                err = MEMKIND_ERROR_RUNTIME;
-            }
-        }
-    }
-    return err;
-}
-
-static int numanode_bandwidth_compare(const void *a, const void *b)
-{
-    /***************************************************************************
-    *  qsort comparison function for numa_node_bandwidth structures.  Sorts in *
-    *  order of bandwidth and then numanode.                                   *
-    ***************************************************************************/
-    int result;
-    struct numanode_bandwidth_t *aa = (struct numanode_bandwidth_t *)(a);
-    struct numanode_bandwidth_t *bb = (struct numanode_bandwidth_t *)(b);
-    result = (aa->bandwidth > bb->bandwidth) - (aa->bandwidth < bb->bandwidth);
-    if (result == 0) {
-        result = (aa->numanode > bb->numanode) - (aa->numanode < bb->numanode);
-    }
-    return result;
 }
 
 MEMKIND_EXPORT void memkind_hbw_init_once(void)
